requirements.txt
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.6
datasets==2.18.0
deep-sort-realtime==1.3.2
dill==0.3.8
docker-pycreds==0.4.0
evaluate==0.4.1
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2023.12.1
furl==2.1.3
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
huggingface-hub==0.22.1
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
lapx==0.5.5
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
opencv-python==4.8.1.78
orderedmultidict==1.0.1
packaging==23.2
pandas==2.0.3
pathlib2==2.3.7.post1
Pillow==10.1.0
pkgutil-resolve-name==1.3.10
protobuf==4.25.2
psutil==5.9.6
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyDrive==1.3.1
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
responses==0.18.0
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
sentry-sdk==1.43.0
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.39.1
triton==2.1.0
typing-extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.230
uritemplate==4.1.1
urllib3==2.1.0
wandb==0.16.5
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
